refactor: seperate OverlayManager from measurement
